backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.1
psycopg2-binary==2.9.9
httpx==0.25.1
pytest==7.4.3
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
transformers==4.35.2
scipy==1.11.4
pillow==10.1.0
matplotlib==3.8.2
seaborn==0.13.0
wfdb==4.1.2
tqdm==4.66.1
tensorflow==2.15.0

# AI Voice Assistant Dependencies
google-generativeai==0.3.2
deep-translator==1.11.4
langdetect==1.0.9
gtts==2.5.0
pydub==0.25.1
SpeechRecognition==3.10.1
python-dotenv==1.0.0

# Authentication & Email
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
aiosmtplib==3.0.1
jinja2==3.1.3

# PDF Generation
reportlab==4.0.7

